+++ CHANGELOG.md
@@ -9,6 +9,8 @@ History
 - Configuration: Allow merging from default values
 - Configuration: Add `insert`/`insert_if_missing` strategies for 
 - Shell: Allow special characters from configuration to be properly exported in shell.
+- Docker: Add `userNameToUid`/`groupNameToGid` jsonnet functions.
+- Docker: Add `docker.user.name`/`docker.user.group` configuration parameters.
 
 1.2.3 (2020-11-13)
 ------------------